+++ src/core/compress/elias_fano_compression.h
@@ -19,7 +19,7 @@ class EliasFanoCompression : public Compression {
   ~EliasFanoCompression();
   void compress(const uint32_t* in, size_t in_size, uint8_t* out,
                 size_t& out_size) override;
-  uint32_t* decompress(const uint8_t* in, size_t in_size, uint32_t* out,
+  uint8_t* decompress(const uint8_t* in, size_t in_size, uint32_t* out,
                        size_t& out_size) override;
 };
